fix (src/controller): begin of gen. fix of stop
This is beginning of rewriting code in stages, silmuntaniously testing it
on real turtlebot. The main goal is to make stop work! On this stage the
code contains TurtleBot class, which can initialize and move linear
with 0.5 m/s velocity. It also contains rightly written
'if __name__ == '__main__':' part. Stop works.