fix (src/slam_explorer): New Frame!
The coordinate system of /map (which was lower right corner of the map,
at the moment, when map was initialized) is now shifted to the lower
right corner (where cell(0,0) is always situated) and now the goal coordinates
can be inputed simply as cell row and column on the map array. It was done like
this: two new reference frames ("lower right corner", "goal coordinates")
were created. And goal coordinates for action lib is inputed as
translation /map->lower right corner->goal.