add (src/slam_explorer/DFD_detector): Adds prohibition of negative gradient!!
Adds prohibition of negative gradient, because it mostly encounters like this:

0 0 0 0 0 0 0 100 100	| 50   	|
0 0 0 0 0 0 0 100 UC 	| 50	|
0 known cells 100 100	| 50	|
0 0 0 0 0 0 0 100 100	| 50	|

The map has layer of known cells, then thin layer of unknown cells and then
layer of wall. So the gradient goes up then stays then goes down. That can cause
the highest gradient, when frontier between unknown cells and wall is closer
to the robot. Also because of the bug i fixed, Magnitude map of gradient stored
smaller gradient value as highest value because of overflowing the datatype.

So now program works much better, robot explores the terrain near him, not
throwing himself to the walls. But still I need to implement:

- list of unreachable frontiers or cells, so the robot doesn't throw
  himself on them over and over

- check if there is wall (occupied cells) near goal, if so -> move the goal
  in oposite direction from the wall