feat (src/slam_explorer/TBSE): Adds finite state machine!!
Adds finite state machine to the main while loop of the
TBSE class. It prevents the slipping of frontiers on my side, when
server holds code, that goal is reached for more than one sample, because
after reaching the result, state machine goes directly to the 1 state instead
of running the whole loop again and again. What I've noticed is that frontier
slipping is still goes in the second state:
"SETTING UP THE GOAL IN THE MAP FRAME". It's because if transformation
is not found, state machine stays in the second state and tries to get
transformation of new frontier instead of waiting for transformation for
old one